Revert "Merge main"
This reverts commit 3030da4838b86cdea6ada8b5a21a06f5d26b7f02.

+++ pkg/security/seclwin/model/args_envs.go
@@ -20,7 +20,7 @@ const (
 
 // ArgsEnvs raw value for args and envs
 type ArgsEnvs struct {
-	ID        uint64
+	ID        uint32
 	Size      uint32
 	ValuesRaw [MaxArgEnvSize]byte
 }
